Fix lua_*upvalue() when upvalue names aren't in debug info
`lua_getupvalue()` and `lua_setupvalue()` don't behave as expected when working with Lua closure whose `Proto` has no debug info. The code currently uses `sizeupvalues` to do bounds checking of upvalue indices, but that's the size of the upvalue _names_ array. It will always be `0` if the `Proto` doesn't have debug info.

This uses `nups` instead, and just returns `""` as the upvalue name if we don't have one, same as for C closures.